requirement.txt
Django==2.2.2
django-allauth==0.40.0
google-api-python-client==1.6.4
oauth2client==4.1.2
jsonpickle==0.9.6
django-widget-tweaks==1.4.5
numpy
python-dateutil==2.8.1
Pillow==6.2.1
#dateutil==2.8.1
